🐛 01ACZ9 task: break knowledge request import cycle
