+++ .agentplane/tasks/202607221852-01ACZ9/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607221852-01ACZ9/serve-bounded-knowledge-requests-during-agent-ep",
   "created_at": "2026-07-30T13:20:41.795Z",
-  "diffstat_sha256": "sha256:5987b4e25613f5611381483048525886fb918642bc1fea397eefc491392428e2",
+  "diffstat_sha256": "sha256:a5aa16be61fc824ba6635cd5d94e5e86a74d93f80a9fc0b60a75ba86a6557be4",
   "last_verified_at": "2026-07-30T15:09:32.150Z",
   "last_verified_diffstat_sha256": "sha256:5987b4e25613f5611381483048525886fb918642bc1fea397eefc491392428e2",
   "pr_number": 4693,
