+++ .agentplane/tasks/202607221852-01ACZ9/pr/review.md
@@ -38,13 +38,14 @@ Created: 2026-07-30T13:20:41.795Z
  .../src/runner/task-state-render-semantic.ts       |   2 +-
  packages/agentplane/src/runner/types/invocation.ts |   9 +
  .../src/runner/usecases/agent-work-order-build.ts  |   3 +
- .../usecases/task-knowledge-request-audit.ts       | 212 ++++++++
- .../usecases/task-knowledge-request-codec.ts       |  56 ++
- .../task-knowledge-request-lifecycle.test.ts       | 322 +++++++++++
+ .../usecases/task-knowledge-request-audit.ts       | 212 +++++++++
+ .../usecases/task-knowledge-request-codec.ts       |  56 +++
+ .../task-knowledge-request-lifecycle.test.ts       | 322 ++++++++++++++
  .../usecases/task-knowledge-request-lifecycle.ts   |  98 ++++
- .../usecases/task-knowledge-request-scope.ts       |  88 +++
- .../runner/usecases/task-knowledge-request.test.ts | 494 +++++++++++++++++
- .../src/runner/usecases/task-knowledge-request.ts  | 597 +++++++++++++++++++++
+ .../usecases/task-knowledge-request-response.ts    | 168 +++++++
+ .../usecases/task-knowledge-request-scope.ts       |  88 ++++
+ .../runner/usecases/task-knowledge-request.test.ts | 494 +++++++++++++++++++++
+ .../src/runner/usecases/task-knowledge-request.ts  | 435 ++++++++++++++++++
  .../runner/usecases/task-knowledge-retrieval.ts    |  14 +
  .../task-run-bootstrap.result-examples.test.ts     |   6 +-
  .../src/runner/usecases/task-run-bootstrap.ts      |   9 +
@@ -54,11 +55,11 @@ Created: 2026-07-30T13:20:41.795Z
  packages/core/src/runner/agent-semantic-result.ts  |  21 +
  .../core/src/runner/agent-work-order-fixtures.ts   |   1 +
  packages/core/src/schemas/index.ts                 |   4 +
- schemas/agent-semantic-result.schema.json          |  87 ++-
+ schemas/agent-semantic-result.schema.json          |  87 +++-
  ...ent-semantic-result-v2.needs-context.valid.json |   7 +-
  .../agent-work-order-v2.camel-case.compat.json     |   1 +
  schemas/examples/agent-work-order-v2.valid.json    |   1 +
- 29 files changed, 2204 insertions(+), 34 deletions(-)
+ 30 files changed, 2210 insertions(+), 34 deletions(-)
 ```
 
 </details>
